Move EventCard into event module
