+++ event/EventCard.tsx
