+++ screens/ProfileScreen/EventHistoryModal.tsx
@@ -1,7 +1,7 @@
 import React, { SetStateAction } from "react"
 import { Title } from "@components/Text"
 import { TouchableIonicon } from "@components/common/Icons"
-import { EventCard } from "@components/EventCard"
+import { EventCard } from "@event/EventCard"
 import { AppStyles } from "@lib/AppColorStyle"
 import { ClientSideEvent } from "@event/ClientSideEvent"
 import { FlatList, ListRenderItemInfo, Modal, StyleSheet } from "react-native"
